Move LRUCache into a separate package
