+++ internal/cache/lru/lru.go
@@ -1,4 +1,4 @@
-package cache
+package lru
 
 import (
 	"hash/maphash"
