+++ internal/cache/lru/lru_test.go
@@ -1,4 +1,4 @@
-package cache
+package lru
 
 import (
 	"bytes"
